requirements.txt
Pillow==10.4.0
numpy==2.0.1
scipy==1.13.1
typer==0.12.3
